requirements.txt
fastapi[all]
psycopg2-binary
sqlalchemy
pydantic
python-jose[cryptography]